[Refactor/SWM-211] 수정, 게시로직 리펙토링, 간단한 나라 API 추가 (#68)
* [SWM-195] feat : Feign Client를 위한 기본세팅

* [SWM-211] refactor : 수정, 게시 로직의 리턴값을 id에서 조회리턴값으로 변경

* [SWM-211] refactor : 성향을 int로 받고, 리턴값도 int로 리턴하도록 리팩토링

* [SWM-217] refactor : 나라명, 대륙명, 나라사진 리턴하는 API 개발

+++ src/main/java/com/example/tripmingle/dto/res/matching/MatchingUserResDTO.java
@@ -6,28 +6,28 @@ import lombok.Getter;
 @Getter
 @Builder
 public class MatchingUserResDTO {
-    //유저정보
-    private Long userId;
-    private String nickName;
-    private String ageRange;
-    private String gender;
-    private String nationality;
-    private String selfIntroduction;
+	//유저정보
+	private Long userId;
+	private String nickName;
+	private String ageRange;
+	private String gender;
+	private String nationality;
+	private String selfIntroduction;
 
-    //유저 성향 정보
-    private Long userPersonalityId;
-    private double vegan; // 비건 여부
-    private double islam; // 이슬람 신앙
-    private double hindu; // 힌두 신앙
-    private double smoking; // 흡연 여부
-    private double budget; // 예산 범위
-    private double accommodationFlexibility; // 숙소 선택 성향
-    private double foodFlexibility; // 음식 선택 성향
-    private double activity; // 선호 활동 - 액티비티
-    private double instagramPicture; // 선호 활동 - 인스타사진
-    private double foodExploration; // 선호 활동 - 맛집탐방
-    private double adventure; // 선호 활동 - 탐험
-    private double personality; // 성격
-    private double schedule; // 일정 계획 성향
+	//유저 성향 정보
+	private Long userPersonalityId;
+	private int vegan; // 비건 여부
+	private int islam; // 이슬람 신앙
+	private int hindu; // 힌두 신앙
+	private int smoking; // 흡연 여부
+	private int budget; // 예산 범위
+	private int accommodationFlexibility; // 숙소 선택 성향
+	private int foodFlexibility; // 음식 선택 성향
+	private int activity; // 선호 활동 - 액티비티
+	private int instagramPicture; // 선호 활동 - 인스타사진
+	private int foodExploration; // 선호 활동 - 맛집탐방
+	private int adventure; // 선호 활동 - 탐험
+	private int personality; // 성격
+	private int schedule; // 일정 계획 성향
 
 }
